% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_matrix.R
\name{compute_percentage_matrix}
\alias{compute_percentage_matrix}
\title{Compute a matrix of mass percentages of nutrients in each fertilizer}
\usage{
compute_percentage_matrix(nutrient_matrix)
}
\arguments{
\item{nutrient_matrix}{A [fertilizer x nutrient] matrix in mmol/mol}
}
\value{
A matrix of nutrient mass percentages per fertilizer
}
\description{
Compute a matrix of mass percentages of nutrients in each fertilizer
}
