% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_canonical_to_unit}
\alias{from_canonical_to_unit}
\title{Convert from canonical unit to output unit}
\usage{
from_canonical_to_unit(vals, unit_out, canonical_unit = "mmol/L")
}
\arguments{
\item{vals}{Numeric vector of values.}

\item{unit_out}{Unit string.}

\item{canonical_unit}{Canonical unit.}
}
\description{
Convert from canonical unit to output unit
}
\keyword{internal}
