% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.R
\name{ion_diffusion_25C}
\alias{ion_diffusion_25C}
\title{Lookup table of diffusion coefficients at 25 °C for EC calculation}
\usage{
ion_diffusion_25C()
}
\value{
A data.frame with columns: ion, z, D_1e5_cm2_s.
}
\description{
Returns a data frame with ions, charge, and diffusion coefficients at 25 °C.
Diffusion values are stored as numbers in 10^-5 cm^2/s (which is numerically
equal to 10^-9 m^2/s).
}
