% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.R
\name{ion_lambda0_25C}
\alias{ion_lambda0_25C}
\title{Lookup table of limiting equivalent conductances at 25 C}
\usage{
ion_lambda0_25C()
}
\value{
A data.frame with columns: ion, z, Lambda0_eq.
}
\description{
Returns a data frame with ions, charge, and limiting equivalent conductances
(Lambda0_eq) at 25 C. Values are expressed as S*cm^2/eq.
}
