% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_nutrient_names}
\alias{normalize_nutrient_names}
\title{Normalize nutrient name keys to match canonical forms}
\usage{
normalize_nutrient_names(x)
}
\arguments{
\item{x}{Character vector of nutrient names.}
}
\description{
Normalize nutrient name keys to match canonical forms
}
\keyword{internal}
