% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_unit}
\alias{normalize_unit}
\title{Normalize unit labels to canonical form}
\usage{
normalize_unit(u, canonical_unit = "mmol/L")
}
\arguments{
\item{u}{Unit string.}

\item{canonical_unit}{The canonical unit to default to.}
}
\description{
Normalize unit labels to canonical form
}
\keyword{internal}
