% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_matrix.R
\docType{data}
\name{percentage_matrix}
\alias{percentage_matrix}
\title{Nutrient mass percentage matrix}
\format{
A numeric matrix with rows as fertilizer formulas and columns as nutrients
(NO3_N, NH4_N, P, K, Ca, Mg, S, Na, Cl, Fe, Mn, Zn, B, Cu, Mo, Si). Each value
is the mass percentage of the nutrient in the fertilizer.
}
\usage{
data(percentage_matrix)
}
\description{
Mass percentages of nutrients in each fertilizer derived from `nutrient_matrix`.
}
