% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb.R
\name{print.bwb_mittelwert}
\alias{print.bwb_mittelwert}
\title{Pretty-print BWB Mittelwerte}
\usage{
\method{print}{bwb_mittelwert}(x, ...)
}
\arguments{
\item{x}{A `bwb_mittelwert` numeric vector.}

\item{...}{Additional arguments passed to `print`.}
}
\value{
The input `bwb_mittelwert` object, invisibly.
}
\description{
Pretty-print BWB Mittelwerte
}
