% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_salt.R
\name{remove_fertilizer}
\alias{remove_fertilizer}
\title{Remove a fertilizer from the nutrient matrix}
\usage{
remove_fertilizer(name, nutrient_matrix)
}
\arguments{
\item{name}{Name of the fertilizer (row name) to remove}

\item{nutrient_matrix}{The current matrix}
}
\value{
Updated matrix
}
\description{
Remove a fertilizer from the nutrient matrix
}
