% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_matrix.R
\name{save_nutrient_matrix}
\alias{save_nutrient_matrix}
\title{Save the nutrient matrix}
\usage{
save_nutrient_matrix(nutrient_matrix, path = "data/nutrient_matrix.rda")
}
\arguments{
\item{nutrient_matrix}{The matrix to save}

\item{path}{File path to save (e.g. "data/nutrient_matrix.rda")}
}
\description{
Save the nutrient matrix
}
