% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_canonical_from_unit}
\alias{to_canonical_from_unit}
\title{Convert from user unit to canonical unit}
\usage{
to_canonical_from_unit(vals, unit_in, canonical_unit = "mmol/L")
}
\arguments{
\item{vals}{Numeric vector of values.}

\item{unit_in}{Unit string.}

\item{canonical_unit}{Canonical unit.}
}
\description{
Convert from user unit to canonical unit
}
\keyword{internal}
