% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_named_numeric}
\alias{to_named_numeric}
\title{Coerce nutrient data into a named numeric vector}
\usage{
to_named_numeric(x)
}
\arguments{
\item{x}{Named numeric vector or data.frame with Nutrient/Achieved columns.}
}
\description{
Coerce nutrient data into a named numeric vector
}
\keyword{internal}
