% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsalt.R
\name{add_fertilizer}
\alias{add_fertilizer}
\title{Add a fertilizer to the nutrient matrix}
\usage{
add_fertilizer(formula, nutrient_matrix)
}
\arguments{
\item{formula}{Chemical formula (e.g. "NaCl", "Ca(NO3)2", "CuSO4·5H2O")}

\item{nutrient_matrix}{Existing nutrient matrix [fertilizer x nutrients]}
}
\value{
Updated nutrient matrix with new fertilizer row
}
\description{
Add a fertilizer to the nutrient matrix
}
