% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.R
\name{ion_diffusion_25C}
\alias{ion_diffusion_25C}
\title{Lookup table of ion diffusion coefficients at 25 C}
\usage{
ion_diffusion_25C()
}
\value{
A data.frame with columns: ion, z, D_1e5_cm2_s.
}
\description{
Returns a data frame with ions, charge, and diffusion coefficients (D) at 25 C.
Values are expressed as 10^-5 cm^2/s.
}
