% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{optimize_nutrients}
\alias{optimize_nutrients}
\title{Optimize Nutrient Solution Using NNLS}
\usage{
optimize_nutrients(nutrient_matrix, target, importance)
}
\arguments{
\item{nutrient_matrix}{A matrix of nutrient contents [nutrient x compound].}

\item{target}{A named numeric vector of target concentrations (mol/L).}

\item{importance}{A named numeric vector (-2 to +2) specifying nutrient importance.}
}
\value{
An object of class 'nutrient_optimization_result'.
}
\description{
Solves for optimal fertilizer amounts to meet target nutrient concentrations.
}
