% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_negative_zero}
\alias{strip_negative_zero}
\title{Remove negative zeros introduced by rounding}
\usage{
strip_negative_zero(x)
}
\arguments{
\item{x}{Numeric vector or matrix.}
}
\description{
Ensures rounded values like -0.00 are displayed as 0.
}
